Move tokenization into preprocessing
